Refactor Gatsby dependency out of Link component
Create a new component for the Design System documentation to link
within the Gatsby site or to use a standard anchor tag for links outside
the site.
