+++ src/components/Link/index.js
@@ -1,24 +1,10 @@
-import { Link as GatsbyLink } from "gatsby";
 import PropTypes from "prop-types";
 import React from "react";
 
 import "./index.css";
 
-const Link = ({ children, inverted, href, activeClassName, ...other }) => {
+const Link = ({ children, inverted, href, ...other }) => {
   const className = `Link ${inverted ? "Link--inverted" : ""}`;
-  // Assume internal links start with exactly one slash
-  if (/^\/(?!\/)/.test(href)) {
-    return (
-      <GatsbyLink
-        className={className}
-        to={href}
-        activeClassName={activeClassName}
-        {...other}
-      >
-        {children}
-      </GatsbyLink>
-    );
-  }
   return (
     <a className={className} href={href} {...other}>
       {children}
@@ -27,7 +13,6 @@ const Link = ({ children, inverted, href, activeClassName, ...other }) => {
 };
 
 Link.propTypes = {
-  activeClassName: PropTypes.string,
   href: PropTypes.string.isRequired,
   inverted: PropTypes.bool,
   children: PropTypes.node.isRequired
