+++ src/pages/assets/generating-sketch-assets.js
@@ -2,7 +2,7 @@ import React from "react";
 
 import Breadcrumbs from "../../components/Breadcrumbs";
 import Cite from "../../components/Cite";
-import Link from "../../components/Link";
+import Link from "../../components/DesignSystemLink";
 import ListItem from "../../components/ListItem";
 import Main from "../../components/Main";
 import NumberedList from "../../components/NumberedList";
