+++ src/pages/assets/images-and-videos.js
@@ -3,7 +3,7 @@ import React from "react";
 import Breadcrumbs from "../../components/Breadcrumbs";
 import BulletedList from "../../components/BulletedList";
 import Cite from "../../components/Cite";
-import Link from "../../components/Link";
+import Link from "../../components/DesignSystemLink";
 import ListItem from "../../components/ListItem";
 import Main from "../../components/Main";
 import Page from "../../components/Page";
