+++ src/pages/assets/web-assets.js
@@ -6,11 +6,11 @@ import BulletedList from "../../components/BulletedList";
 import Content from "../../components/Content";
 import Contents from "../../components/Contents";
 import ContentsMenu from "../../components/ContentsMenu";
+import Link from "../../components/DesignSystemLink";
 import Figure from "../../components/Figure";
 import Figures from "../../components/Figures";
 import Heading from "../../components/Heading";
 import Image from "../../components/Image";
-import Link from "../../components/Link";
 import ListItem from "../../components/ListItem";
 import Main from "../../components/Main";
 import Page from "../../components/Page";
