+++ src/pages/components/content.js
@@ -11,13 +11,13 @@ import Code from "../../components/Code";
 import Content from "../../components/Content";
 import Contents from "../../components/Contents";
 import ContentsMenu from "../../components/ContentsMenu";
+import Link from "../../components/DesignSystemLink";
 import Details from "../../components/Details";
 import Example from "../../components/Example";
 import Fact from "../../components/Fact";
 import Heading from "../../components/Heading";
 import Image from "../../components/Image";
 import Lede from "../../components/Lede";
-import Link from "../../components/Link";
 import ListItem from "../../components/ListItem";
 import Main from "../../components/Main";
 import NumberedList from "../../components/NumberedList";
