+++ src/pages/components/forms.js
@@ -12,12 +12,12 @@ import Contents from "../../components/Contents";
 import ContentsMenu from "../../components/ContentsMenu";
 import CurrencyInput from "../../components/CurrencyInput";
 import DateInput from "../../components/DateInput";
+import Link from "../../components/DesignSystemLink";
 import ErrorSummary from "../../components/ErrorSummary";
 import Example from "../../components/Example";
 import Figure from "../../components/Figure";
 import Figures from "../../components/Figures";
 import Heading from "../../components/Heading";
-import Link from "../../components/Link";
 import ListItem from "../../components/ListItem";
 import Main from "../../components/Main";
 import Page from "../../components/Page";
