+++ src/pages/components/navigation.js
@@ -8,11 +8,11 @@ import Cite from "../../components/Cite";
 import Content from "../../components/Content";
 import Contents from "../../components/Contents";
 import ContentsMenu from "../../components/ContentsMenu";
+import Link from "../../components/DesignSystemLink";
 import DonateLink from "../../components/DonateLink";
 import DownloadLink from "../../components/DownloadLink";
 import Example from "../../components/Example";
 import Heading from "../../components/Heading";
-import Link from "../../components/Link";
 import ListItem from "../../components/ListItem";
 import Main from "../../components/Main";
 import Page from "../../components/Page";
