+++ src/pages/components/structure.js
@@ -7,13 +7,13 @@ import Code from "../../components/Code";
 import Content from "../../components/Content";
 import Contents from "../../components/Contents";
 import ContentsMenu from "../../components/ContentsMenu";
+import Link from "../../components/DesignSystemLink";
 import Example from "../../components/Example";
 import Figure from "../../components/Figure";
 import Figures from "../../components/Figures";
 import Footer from "../../components/Footer";
 import Header from "../../components/Header";
 import Heading from "../../components/Heading";
-import Link from "../../components/Link";
 import ListItem from "../../components/ListItem";
 import Main from "../../components/Main";
 import Menu from "../../components/Menu";
