+++ src/pages/patterns/filtering.js
@@ -3,10 +3,10 @@ import React from "react";
 import Breadcrumbs from "../../components/Breadcrumbs";
 import BulletedList from "../../components/BulletedList";
 import Cite from "../../components/Cite";
+import Link from "../../components/DesignSystemLink";
 import Example from "../../components/Example";
 import Filter from "../../components/Filter";
 import Form from "../../components/Form";
-import Link from "../../components/Link";
 import ListItem from "../../components/ListItem";
 import Main from "../../components/Main";
 import Page from "../../components/Page";
