+++ src/pages/patterns/interacting-with-young-people.js
@@ -7,12 +7,12 @@ import Cite from "../../components/Cite";
 import Content from "../../components/Content";
 import Contents from "../../components/Contents";
 import ContentsMenu from "../../components/ContentsMenu";
+import Link from "../../components/DesignSystemLink";
 import Figure from "../../components/Figure";
 import Figures from "../../components/Figures";
 import Heading from "../../components/Heading";
 import Image from "../../components/Image";
 import Lede from "../../components/Lede";
-import Link from "../../components/Link";
 import ListItem from "../../components/ListItem";
 import Main from "../../components/Main";
 import Page from "../../components/Page";
