+++ src/pages/patterns/start-page.js
@@ -3,9 +3,9 @@ import React from "react";
 import Breadcrumbs from "../../components/Breadcrumbs";
 import BulletedList from "../../components/BulletedList";
 import Cite from "../../components/Cite";
+import Link from "../../components/DesignSystemLink";
 import Figure from "../../components/Figure";
 import Figures from "../../components/Figures";
-import Link from "../../components/Link";
 import ListItem from "../../components/ListItem";
 import Main from "../../components/Main";
 import Page from "../../components/Page";
