+++ src/pages/standards/colour.js
@@ -8,10 +8,10 @@ import ColourGreyFigure from "../../components/ColourGreyFigure";
 import Content from "../../components/Content";
 import Contents from "../../components/Contents";
 import ContentsMenu from "../../components/ContentsMenu";
+import Link from "../../components/DesignSystemLink";
 import Figure from "../../components/Figure";
 import Figures from "../../components/Figures";
 import Heading from "../../components/Heading";
-import Link from "../../components/Link";
 import ListItem from "../../components/ListItem";
 import Main from "../../components/Main";
 import Page from "../../components/Page";
