+++ src/pages/standards/typography.js
@@ -7,9 +7,9 @@ import Code from "../../components/Code";
 import Content from "../../components/Content";
 import Contents from "../../components/Contents";
 import ContentsMenu from "../../components/ContentsMenu";
+import Link from "../../components/DesignSystemLink";
 import Figures from "../../components/Figures";
 import Heading from "../../components/Heading";
-import Link from "../../components/Link";
 import ListItem from "../../components/ListItem";
 import Main from "../../components/Main";
 import Page from "../../components/Page";
